# public/swagger.yml
...
swagger: '2.0'
...
info:
  version: "1"
  title: Clowder API
  description: |
    Documentation for the [Clowder](http://clowder.ncsa.illinois.edu/) API

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: clowder.ncsa.illinois.edu
basePath: /clowder/api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
securityDefinitions:
  login:
    type: basic
  key:
    type: apiKey
    name: key
    in: query
security:
  - key: []
  - login: []

################################################################################
# Parameters                                                                   #
################################################################################

################################################################################
# Paths                                                                        #
################################################################################
paths:
...
  /files/{id}/metadata:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
      tags:
        - files
      summary: Add technical metadata to file
      description: |
        Metadata in attached JSON object will describe the file's described
        resource, not the file object itself.
    get:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
      tags:
        - files
      summary: Retrieve physical file object metadata
      description: |
        Get metadata of the file object (not the resource it describes) as JSON.
        For example, size of file, date created, content type, filename.
...
  /files/{id}.{format}metadat.{format}:
    get:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
        - required: true
          type: string
          name: format
          in: path
      tags:
        - files
      summary: |
        Get metadata of the resource described by the file that were input as
        XML
      description: A file is the raw bytes plus metadata.
...
  /datasets/{id}/metadata:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
      tags:
        - datasets
      summary: Add metadata to dataset
      description: Returns success of failure
    get:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
      tags:
        - datasets
      summary: |
        Retrieve available metadata definitions for a dataset. It is an
        aggregation of the metadata that a space belongs to.
      description: A dataset is a container for files and metadata
...
  /extractions/{id}/metadata:
    get:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
      tags:
        - extractions
      summary: Provides the metadata extracted from the file
      description: Retruns Status of extractions and metadata extracted so far
...